Functional FLASH read/update via USART

+++ rc5.h
@@ -11,7 +11,7 @@ void rc5_mode_1();
 void rc5_mode_2();
 void rc5_serial_mode();
 void rc5_config_mode();
-void rc5_config_write(uint8_t addr, uint8_t cmd, uint8_t value);
+void rc5_config_write(uint8_t addr, uint8_t cmd, uint16_t value);
 
 void get_remote_data();
 void post_remote_data();
